chore: IAC for new lambda, slight logic tweak & first test
